consolidate package under src
