--- dji_metadata_embedder/__init__.py
@@ -1,10 +1,0 @@
-from .embedder import DJIMetadataEmbedder
-from .per_frame_embedder import embed_flight_path, extract_frame_locations
-from .dat_parser import parse_v13 as parse_dat_v13
-
-__all__ = [
-    "DJIMetadataEmbedder",
-    "embed_flight_path",
-    "extract_frame_locations",
-    "parse_dat_v13",
-]
