+++ pyproject.toml
@@ -26,7 +26,6 @@ keywords = ["dji", "drone", "metadata", "video", "gps", "telemetry", "srt"]
 dependencies = [
     "click>=8.0.0",
     "rich>=10.0.0",
-    "pathlib",
 ]
 dynamic = ["version"]
