+++ src/dji_metadata_embedder/__init__.py
@@ -1,4 +1,23 @@
-﻿"""DJI Drone Metadata Embedder."""
+"""DJI Drone Metadata Embedder."""
 __version__ = "1.0.3"
-from .cli import main
-__all__ = ["main", "__version__"]
+
+# Import check to ensure files were moved correctly
+try:
+    from .embedder import DJIMetadataEmbedder
+    from .per_frame_embedder import embed_flight_path, extract_frame_locations
+    from .dat_parser import parse_v13 as parse_dat_v13
+    from .cli import main
+
+    __all__ = [
+        "__version__",
+        "main",
+        "DJIMetadataEmbedder",
+        "embed_flight_path",
+        "extract_frame_locations",
+        "parse_dat_v13",
+    ]
+except ImportError as e:  # noqa: BLE001
+    import warnings
+
+    warnings.warn(f"Some modules could not be imported: {e}")
+    __all__ = ["__version__"]
