+++ src/dji_metadata_embedder/core/processor.py
