+++ src/dji_metadata_embedder/core/validator.py
