+++ src/dji_metadata_embedder/dat_parser.py
