+++ src/dji_metadata_embedder/embedder.py
@@ -10,8 +10,8 @@ from rich.progress import Progress
 
 from .dat_parser import parse_v13 as parse_dat_v13
 from .utilities import apply_redaction, setup_logging
-from src.utils.dependency_manager import DependencyManager
-from src.utils import system_info
+from .utils.dependency_manager import DependencyManager
+from .utils import system_info
 
 logger = logging.getLogger(__name__)
