+++ src/dji_metadata_embedder/metadata_check.py
