+++ src/dji_metadata_embedder/per_frame_embedder.py
