+++ src/dji_metadata_embedder/telemetry_converter.py
