+++ src/dji_metadata_embedder/utilities.py
