+++ src/dji_metadata_embedder/utils/dependency_manager.py
