+++ src/dji_metadata_embedder/utils/system_info.py
