+++ tests/test_format3_parsing.py
@@ -2,8 +2,6 @@ import tempfile
 from pathlib import Path
 import sys
 
-sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
-
 from dji_metadata_embedder import DJIMetadataEmbedder
 from dji_metadata_embedder import telemetry_converter
