+++ tests/test_parsing.py
@@ -2,7 +2,6 @@ import sys
 from pathlib import Path
 import subprocess
 
-sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 from dji_metadata_embedder import DJIMetadataEmbedder
