+++ tests/test_per_frame_embedder.py
@@ -2,8 +2,6 @@ import sys
 from pathlib import Path
 import subprocess
 
-sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
-
 from dji_metadata_embedder import embed_flight_path, extract_frame_locations
