+++ tests/test_system_info.py
@@ -1,9 +1,6 @@
 import sys
-from pathlib import Path
 
-sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
-
-from src.utils import system_info
+from dji_metadata_embedder.utils import system_info
 
 
 def test_python_info():
